Fix: parsing strings containing large numbers
There was a bug that would crash (overflow exception) serialization
when serializing strings that contain large numbers, for instance
"999999999876", which is a legal string, but for some reason
serializer wants to temporarilly transfor it to an int - which causes
an expected exception, as it is too big for an int.

To avoid this, added limit of how many digits should be
considered numbers.

Signed-off-by: zeljko <zeljko@zwr.fi>